requirements.txt
bandit==1.6.2
gitdb==4.0.5
GitPython==3.1.11
mypy==0.790
mypy-extensions==0.4.3
pbr==5.5.1
PyYAML==5.3.1
six==1.15.0
smmap==3.0.4
stevedore==3.2.2
tabulate==0.8.7
typed-ast==1.4.1
typing-extensions==3.7.4.3
